+++ appTools/ToolDrilling.py
@@ -489,6 +489,7 @@ class ToolDrilling(AppTool, Excellon):
             self.ui.level.setChecked(True)
         else:
             self.ui.level.setChecked(False)
+        self.on_level_changed(self.ui.level.isChecked())
 
     def on_level_changed(self, checked):