+++ appTools/ToolMilling.py
@@ -522,6 +522,7 @@ class ToolMilling(AppTool, Excellon):
             self.ui.level.setChecked(True)
         else:
             self.ui.level.setChecked(False)
+        self.on_level_changed(self.ui.level.isChecked())
 
     def on_level_changed(self, checked):