+++ appTools/ToolNCC.py
@@ -645,6 +645,7 @@ class NonCopperClear(AppTool, Gerber):
             self.ui.level.setChecked(True)
         else:
             self.ui.level.setChecked(False)
+        self.on_level_changed(self.ui.level.isChecked())
 
     def on_level_changed(self, checked):
         if not checked: