+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG for FlatCAM beta
 
 - NCC Tool - remade the 'Beginner/Advanced' Mode
 - Paint Tool - remade the 'Beginner/Advanced' Mode
+- Cutout Tool - remade the 'Beginner/Advanced' Mode
+- minor changes to the UI to all Tools with 'Beginner/Advanced' Mode
 
 9.11.2020